+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Line wrap the file at 100 chars.                                              Th
 - Add in-app updates to Windows and macOS. This new feature lets you download, verify, and install
   new versions from within the app.
 
+#### Linux
+- Make it possible to run the app with cgroups v1 fully disabled. Note that split tunneling is
+  currently unavailable when this is the case.
+
 ### Changed
 #### macOS
 - Use a local DNS resolver on the 127.0.0.0/8 network, regardless of macOS version.
@@ -35,7 +39,6 @@ Line wrap the file at 100 chars.                                              Th
 - Add grace period when best default route goes away to reduce frequency of random reconnects.
 
 ### Security
-#### Windows
 #### Windows
 - Enable control flow integrity checks (CFG) for some C++ code. This excludes `wintun`,
   `wireguard-nt`, and OpenVPN. This addresses `MLLVD-CR-24-101` to the extent that we found